Move components from src/components/pages to modules
