+++ src/components/layout/dashboard/contextHeader/useDashboardBreadcrumbs.tsx
@@ -3,9 +3,9 @@ import { merge, startCase } from 'lodash-es';
 import { useMemo } from 'react';
 import { useLocation } from 'react-router-dom';
 
-import { ClientDashboardContext } from '@/components/pages/ClientDashboard';
-import { EnrollmentDashboardContext } from '@/components/pages/EnrollmentDashboard';
 import useCurrentPath from '@/hooks/useCurrentPath';
+import { ClientDashboardContext } from '@/modules/client/components/pages/ClientDashboard';
+import { EnrollmentDashboardContext } from '@/modules/enrollment/components/pages/EnrollmentDashboard';
 import { clientBriefName, enrollmentName } from '@/modules/hmis/hmisUtil';
 import { ProjectDashboardContext } from '@/modules/projects/components/ProjectDashboard';
 import {
