+++ src/modules/enrollment/components/pages/EnrollmentDashboard.tsx
@@ -1,13 +1,13 @@
 import { Container } from '@mui/material';
 import { isNil } from 'lodash-es';
 import { useMemo, useState } from 'react';
-import { Outlet, useOutletContext } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 
-import Loading from '../elements/Loading';
-import ContextHeaderContent from '../layout/dashboard/contextHeader/ContextHeaderContent';
-import DashboardContentContainer from '../layout/dashboard/DashboardContentContainer';
-import SideNavMenu from '../layout/dashboard/sideNav/SideNavMenu';
-import NotFound from './NotFound';
+import Loading from '../../../../components/elements/Loading';
+import ContextHeaderContent from '../../../../components/layout/dashboard/contextHeader/ContextHeaderContent';
+import DashboardContentContainer from '../../../../components/layout/dashboard/DashboardContentContainer';
+import SideNavMenu from '../../../../components/layout/dashboard/sideNav/SideNavMenu';
+import NotFound from '../../../../components/pages/NotFound';
 
 import {
   useDashboardBreadcrumbs,
@@ -141,8 +141,5 @@ export function isEnrollmentDashboardContext(
     !!value.hasOwnProperty('enrollment')
   );
 }
-
-export const useEnrollmentDashboardContext = () =>
-  useOutletContext<EnrollmentDashboardContext>();
 
 export default EnrollmentDashboard;
