+++ src/modules/enrollment/hooks/useEnrollmentDashboardContext.tsx
@@ -1,0 +1,6 @@
+import { useOutletContext } from 'react-router-dom';
+import { EnrollmentDashboardContext } from '@/modules/enrollment/components/pages/EnrollmentDashboard';
+
+export default function useEnrollmentDashboardContext() {
+  return useOutletContext<EnrollmentDashboardContext>();
+}
