+++ src/modules/household/components/EditHouseholdPage.tsx
@@ -1,8 +1,8 @@
 import BackButtonLink from '@/components/elements/BackButtonLink';
 import PageTitle from '@/components/layout/PageTitle';
-import { useEnrollmentDashboardContext } from '@/components/pages/EnrollmentDashboard';
 import NotFound from '@/components/pages/NotFound';
 import useSafeParams from '@/hooks/useSafeParams';
+import useEnrollmentDashboardContext from '@/modules/enrollment/hooks/useEnrollmentDashboardContext';
 import ManageHousehold from '@/modules/household/components/ManageHousehold';
 import { EnrollmentDashboardRoutes } from '@/routes/routes';
 import { generateSafePath } from '@/utils/pathEncoding';
