+++ src/modules/projects/components/tables/ProjectServicesTable.tsx
@@ -3,7 +3,7 @@ import { useCallback, useMemo } from 'react';
 import { ColumnDef } from '@/components/elements/table/types';
 import ClientName from '@/modules/client/components/ClientName';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
-import { SERVICE_BASIC_COLUMNS } from '@/modules/enrollment/components/dashboardPages/EnrollmentServicesPage';
+import { SERVICE_BASIC_COLUMNS } from '@/modules/enrollment/components/pages/EnrollmentServicesPage';
 import EnrollmentDateRangeWithStatus from '@/modules/hmis/components/EnrollmentDateRangeWithStatus';
 import { useFilters } from '@/modules/hmis/filterUtil';
 import { EnrollmentDashboardRoutes } from '@/routes/routes';
