+++ src/modules/scanCards/components/ClientScanCards.tsx
@@ -7,7 +7,7 @@ import {
 } from './ManageCardButtons';
 import { ColumnDef } from '@/components/elements/table/types';
 import PageTitle from '@/components/layout/PageTitle';
-import { useClientDashboardContext } from '@/components/pages/ClientDashboard';
+import useClientDashboardContext from '@/modules/client/hooks/useClientDashboardContext';
 import { lastUpdatedBy } from '@/modules/hmis/hmisUtil';
 import {
   GetClientScanCardCodesDocument,
