+++ src/modules/services/components/ClientServices.tsx
@@ -9,7 +9,7 @@ import GenericTableWithData from '@/modules/dataFetching/components/GenericTable
 import {
   SERVICE_BASIC_COLUMNS,
   SERVICE_COLUMNS,
-} from '@/modules/enrollment/components/dashboardPages/EnrollmentServicesPage';
+} from '@/modules/enrollment/components/pages/EnrollmentServicesPage';
 import EnrollmentDateRangeWithStatus from '@/modules/hmis/components/EnrollmentDateRangeWithStatus';
 import { useFilters } from '@/modules/hmis/filterUtil';
 import { EnrollmentDashboardRoutes } from '@/routes/routes';
