+++ tsconfig.json
@@ -17,6 +17,7 @@
     "jsx": "react-jsx",
     "types": ["node", "vite/client"],
     "paths": {
+      "@": ["./src"],
       "@/*": ["./src/*"]
     }
   },
